Finish Header, changed logo and picture
I changed the logo and the picture in the header section, edited the
styling of the remaining header elements, such as the button, h1 and
main subtext, and finished the footer. I also removed a class name from
the index page, as it was unnecessary, and removed the old images I used
for the logo and header image. I still need to replace the placeholder
text.